refactor(iris): extract priority scoring from ContextStore
Move PRIORITY/TIME_WINDOW weight tables, MS_PER_DAY/MS_PER_HOUR units,
the two list comparators, and calculateExercisePriority /
calculateCoursePriority into a new pure module
src/extension/services/iris/contextPriorityScorer.ts.

The calculate functions become free functions (they only depended on
their argument and Date.now()). PRIORITY and TIME_WINDOW stay
module-internal. ContextStore drops from 611 to 498 LOC and now imports
the four public functions.

No behavior change. ContextStore test suite (incl. the 3 Task 0
safety-net tests) still green.
